PRISM-233 | Adding EnvAdapter component to emit custom events containing changed properties.
This was put in a component so it could be selectively included in different facets depending on whether we thought it was necessary. At the moment it's only emitting redux state changes. Event names are bound to redux state names.

Breaking embed script out into two parts so the JS embed portion can be invoked separately from the style embed.

A lot of configuration streamlining and cleanup in order to change swatch buttons based on embed location props.

Intentionally excluding snapshots and related test updates.