PRISM-307 | Removing main.css. This is a lot of changes, but they're really pretty minor.
In an effort to produce Facet-specific CSS files, all the utility-esque stuff that used to be manually included in main.scss and piped through index.jsx has been broken out into more modular SCSS includes so that each utility can be imported into the files where they're used. All CSS (and JS, for that matter) utilities should be done in this style going forward to assist in discoverability and modularity.