PRISM-377 | Resolving IE11 issues related to ES6 syntax and dependency polyfilling. Involves using the -cjs versions of react-dnd dependencies.
IMPORTANT: This GLOBALLY polyfills Symbols, which I do not want and would like to change. We should not pollute the global namespace other than the PRISM object.

The original intent was to resolve this issue with tooling. This polyfilling SHOULD be able to be done automatically, but it was not for reasons I (Cody) was not able to weed out. As a result, I have created a facetPolyfills.js file which is included with all the other facetBinder stuff. In here I'm manually including core-js Symbols polyfills, and in the babel config I'm specifically EXcluding those so they're not included twice.