PRISM-506 | Correcting issue where selected and focused colors in color wall were not being announced.
Initially this was a very simple fix -- just a matter of changing a null check in the getAudioMessaging method of ColorWallSwatchList. Then an issue with aria live regions was discovered that required the message to be cleared between updates or else the previous message would never be re-read. To solve this at an application scope, I implemented a facet-wide context version of the announcePolite and announceAssertive methods which feature built-in clearing timeouts. This issue is a screen reader (maybe just VoiceOver?) issue, NOT a react-aria-live issue.