PRISM-1092 | Injecting fallback values for CSS custom props for IE11.
Adding custom postcss plugin to inject fallback values for usages of var() in CSS which map to theme color values, or which have a fallback value in the var().
Updating browserslist to use BL's defaults value, as well as explicitly calling out IE >= 11 and Edge >= 16. This saves a ton of space in our styles since it no longer includes IE 10 support.